APIの作成

+++ db/migrate/20230312075621_create_records.rb
@@ -2,7 +2,7 @@ class CreateRecords < ActiveRecord::Migration[6.1]
   def change
     create_table :records do |t|
       t.string :title
-      t.decimal :study_hour
+      t.float :study_hour
 
       t.timestamps
     end
